refactor: rename pkg/ to twinkit/ across entire codebase
- Rename pkg/ directory to twinkit/
- Update twinkit/go.mod module path to wondertwin/twinkit
- Update go.work to reference ./twinkit
- Update all 6 twin go.mod files (replace directives)
- Update all Go import paths from wondertwin/pkg to wondertwin/twinkit

All tests pass, all builds succeed.
