+++ twin-logodev/cmd/twin-logodev/main.go
@@ -4,8 +4,8 @@ import (
 	"log"
 	"os"
 
-	"github.com/wondertwin-ai/wondertwin/pkg/admin"
-	"github.com/wondertwin-ai/wondertwin/pkg/twincore"
+	"github.com/wondertwin-ai/wondertwin/twinkit/admin"
+	"github.com/wondertwin-ai/wondertwin/twinkit/twincore"
 	"github.com/wondertwin-ai/wondertwin/twin-logodev/internal/api"
 	"github.com/wondertwin-ai/wondertwin/twin-logodev/internal/store"
 )
