+++ twin-resend/internal/store/memory.go
@@ -3,7 +3,7 @@ package store
 import (
 	"encoding/json"
 
-	pkgstore "github.com/wondertwin-ai/wondertwin/pkg/store"
+	pkgstore "github.com/wondertwin-ai/wondertwin/twinkit/store"
 )
 
 // MemoryStore holds all Resend twin state in memory.
