+++ twin-stripe/go.mod
@@ -2,9 +2,9 @@ module github.com/wondertwin-ai/wondertwin/twin-stripe
 
 go 1.25.7
 
-replace github.com/wondertwin-ai/wondertwin/pkg => ../pkg
+replace github.com/wondertwin-ai/wondertwin/twinkit => ../twinkit
 
 require (
 	github.com/go-chi/chi/v5 v5.2.5
-	github.com/wondertwin-ai/wondertwin/pkg v0.0.0-00010101000000-000000000000
+	github.com/wondertwin-ai/wondertwin/twinkit v0.0.0-00010101000000-000000000000
 )
