+++ twin-stripe/internal/store/memory.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"sync"
 
-	pkgstore "github.com/wondertwin-ai/wondertwin/pkg/store"
+	pkgstore "github.com/wondertwin-ai/wondertwin/twinkit/store"
 )
 
 // MemoryStore holds all Stripe twin state in memory.
