+++ twin-twilio/cmd/twin-twilio/main.go
@@ -10,8 +10,8 @@ import (
 	"log"
 	"os"
 
-	"github.com/wondertwin-ai/wondertwin/pkg/admin"
-	"github.com/wondertwin-ai/wondertwin/pkg/twincore"
+	"github.com/wondertwin-ai/wondertwin/twinkit/admin"
+	"github.com/wondertwin-ai/wondertwin/twinkit/twincore"
 	"github.com/wondertwin-ai/wondertwin/twin-twilio/internal/api"
 	"github.com/wondertwin-ai/wondertwin/twin-twilio/internal/store"
 )
