+++ twin-twilio/go.mod
@@ -4,7 +4,7 @@ go 1.25.7
 
 require (
 	github.com/go-chi/chi/v5 v5.2.5
-	github.com/wondertwin-ai/wondertwin/pkg v0.0.0
+	github.com/wondertwin-ai/wondertwin/twinkit v0.0.0
 )
 
-replace github.com/wondertwin-ai/wondertwin/pkg => ../pkg
+replace github.com/wondertwin-ai/wondertwin/twinkit => ../twinkit
